AegisWeb/requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
blinker==1.9.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
dotenv==0.9.9
en_core_web_sm==3.8.0
ffmpeg==1.4
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gspread==6.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
openai==1.68.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0.1
preshed==3.0.9
proto-plus==1.26.1
protobuf==6.30.1
pyahocorasick==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==78.0.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
textsearch==0.0.24
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
